# protocol/patricia: only hash nodes on demand