@jbowens 2016-11-23
lgtm

```
benchmark              old ns/op     new ns/op     delta
BenchmarkInserts-4     142224708     36920148      -74.04%

benchmark              old allocs     new allocs     delta
BenchmarkInserts-4     283501         283499         -0.00%

benchmark              old bytes     new bytes     delta
BenchmarkInserts-4     17648974      15540852      -11.94%
```

👌